feat: add nicer verify message
